[store] refactor: move trait RaftStateSerde to SledSerde and make it public for other component that needs to access sled db
